+++ src/test/scala/edu/ie3/simona/model/em/EmAggregateSimpleSumSpec.scala
@@ -7,27 +7,23 @@
 package edu.ie3.simona.model.em
 
 import edu.ie3.datamodel.models.input.system.SystemParticipantInput
-import edu.ie3.simona.ontology.messages.flex.MinMaxFlexibilityMessage.ProvideMinMaxFlexOptions
+import edu.ie3.simona.ontology.messages.flex.MinMaxFlexOptions
 import edu.ie3.simona.test.common.UnitSpec
 import org.scalatestplus.mockito.MockitoSugar
 import squants.energy.Kilowatts
-
-import java.util.UUID
 
 class EmAggregateSimpleSumSpec extends UnitSpec with MockitoSugar {
 
   "The simple sum aggregating strategy" should {
 
     "calculate ref, min and max power correctly" in {
-      val flexOptions1 = ProvideMinMaxFlexOptions(
-        modelUuid = UUID.randomUUID(),
+      val flexOptions1 = MinMaxFlexOptions(
         ref = Kilowatts(6.0),
         min = Kilowatts(4.0),
         max = Kilowatts(12.0),
       )
 
-      val flexOptions2 = ProvideMinMaxFlexOptions(
-        modelUuid = UUID.randomUUID(),
+      val flexOptions2 = MinMaxFlexOptions(
         ref = Kilowatts(2.0),
         min = Kilowatts(-2.0),
         max = Kilowatts(2.0),
@@ -40,10 +36,10 @@ class EmAggregateSimpleSumSpec extends UnitSpec with MockitoSugar {
         )
       )
 
-      actualResult shouldBe (
+      actualResult shouldBe MinMaxFlexOptions(
         Kilowatts(8.0),
         Kilowatts(2.0),
-        Kilowatts(14.0)
+        Kilowatts(14.0),
       )
     }